% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repro_chunk.R
\name{repro_chunk}
\alias{repro_chunk}
\title{Reproduce Code Chunk}
\usage{
repro_chunk(x, ..., env = rlang::caller_env())
}
\arguments{
\item{x}{\code{\link[shiny]{reactive}} object to make reproducible}

\item{...}{Additional arguments to pass to other methods}

\item{env}{The environment \code{x} is defined in. By default it is the environment of where \code{repro} is called}
}
\value{
A \code{\link{Repro}} object containing all the necessary code and packages to recreate
the provided expression when evaluated.
}
\description{
A short description...
}
\details{
Whilst a default is provided to \code{env}, it is unlikely that this is the same environment \code{x} is defined
in. This is more of a placeholder for sending the correct environment to
}
\keyword{internal}
